refactor(core): promote the runtime-agnostic reconcile machinery
The Docker backend's pure modules are backend-agnostic in everything but
their home. Moving them into rise-backend-core lets the upcoming ECS backend
share one implementation instead of a copy that silently drifts — the Traefik
label vocabulary in particular, where a divergence would be a routing bug
nobody notices until traffic stops.

Moved wholesale: env (merge + length-prefixed env hashing), diff (identity
tuple, drift signals, protected-deployment GC), rolling (throttle, rotation
and readiness verdicts), pod_status, labels, traefik_api, and the shared test
fixtures. Split out of container_builder: the DesiredContainer/DesiredRoute
shape (desired.rs), the name derivations including the collision-resistant
group_service_base (naming.rs), and the Traefik label renderer with its
fail-closed routes_withheld predicate (traefik_render.rs). health.rs keeps its
reqwest-shaped error formatter; the probe policy moves.

Two provider-specific bits had to become parameters rather than constants:

- TraefikApiClient hardcoded the `@docker` suffix when querying serverStatus.
  It now takes the provider name, because the ECS provider registers services
  as `@ecs` and querying the wrong one 404s — which would silently mean no
  readiness signal and a deployment that never becomes Healthy.
- `traefik.docker.network` has no ECS analogue (tasks are on awsvpc ENIs), so
  TraefikRoute::network is now Option and the label is omitted when absent.

Behaviour is preserved and the test accounting proves it: all 146 Docker tests
still pass, unchanged — 82 travelled with the code they cover (env 7, labels
17, diff 19, rolling 22, pod_status 5, traefik_api 11, health 1) and 64 stayed
with build_container, now exercising the moved code through re-exports. Every
existing `rise_backend_docker::{diff,env,labels,…}` path still resolves.
