+++ crates/rise-backend-core/src/lib.rs
@@ -1,35 +1,66 @@
 //! Shared deployment-backend contracts for Rise.
 //!
 //! This crate is the dependency seam between `rise-deploy` and the deployment
-//! runtimes (Kubernetes, Docker). It holds the data models the backends operate
-//! on, the `DeploymentBackend` trait, the provider traits (registry/encryption),
-//! the pure `quantity` and `state_machine` helpers, and the `DeploymentStore`
-//! trait — the database boundary implemented in `rise-deploy` over SQLX.
+//! runtimes (Kubernetes, Docker, ECS). It holds the data models the backends
+//! operate on, the `DeploymentBackend` trait, the provider traits
+//! (registry/encryption), the pure `quantity` and `state_machine` helpers, the
+//! `DeploymentStore` trait — the database boundary implemented in `rise-deploy`
+//! over SQLX — and the runtime-agnostic reconcile machinery every backend
+//! shares: desired-state shaping, env merging/hashing, deterministic naming,
+//! Traefik label rendering, drift diffing, rollout gating and pod-status
+//! reporting.
 //!
 //! It never opens a database connection or runs a query; `sqlx` is a dependency
 //! only for the `FromRow`/`Type` derives on the moved models.
 
 pub mod backend;
 pub mod custom_domain;
+pub mod desired;
+pub mod diff;
+pub mod env;
 pub mod group;
+pub mod health_path;
 pub mod identity;
+pub mod labels;
 pub mod models;
+pub mod naming;
+pub mod pod_status;
 pub mod providers;
 pub mod quantity;
+pub mod rolling;
 pub mod runtime;
 pub mod state_machine;
 pub mod store;
 pub mod system_env;
 pub mod token_ttl;
+pub mod traefik_api;
+pub mod traefik_render;
 pub mod url_builder;
+
+pub mod test_helpers;
 
 pub use backend::{DeploymentBackend, DeploymentUrls};
 pub use custom_domain::validate_custom_domain;
+pub use desired::{DesiredContainer, DesiredRoute};
+pub use diff::{
+    diff_desired_vs_actual, identity_key, spec_key, ActualContainer, InspectedContainer,
+    ReconcileAction,
+};
+pub use env::{hash_env, merge_container_env, pin_system_env, upsert_env};
 pub use group::{normalize_deployment_group, DEFAULT_DEPLOYMENT_GROUP};
+pub use health_path::effective_health_path;
+pub use naming::{
+    container_name, group_app_name, group_service_base, group_service_name, stable_identity_name,
+};
+pub use pod_status::{build_controller_metadata, build_pod_status};
 pub use providers::{
     EncryptionProvider, ImageTagType, RegistryAuthMethod, RegistryCredentials, RegistryProvider,
 };
 pub use rise_deployment_spec::AccessRequirement;
+pub use rolling::{
+    filter_rolling_actions, replica_ready, rolling_rotation_decision, ReadyVerdict,
+    RotationDecision,
+};
 pub use runtime::{
     effective_access_requirement, resolve_deployment_env_vars, resolve_runtime_containers,
     should_have_infrastructure, ResolvedDeploymentEnvVars, DEPLOYING_TIMEOUT_MINUTES,
@@ -38,4 +69,8 @@ pub use runtime::{
 pub use store::DeploymentStore;
 pub use system_env::rise_system_env_vars;
 pub use token_ttl::{refresh_due_after_secs, remint_after_secs};
+pub use traefik_api::TraefikApiClient;
+pub use traefik_render::{
+    render_traefik_labels_for, route_hash_for, routes_withheld, TraefikRenderConfig,
+};
 pub use url_builder::{DeploymentUrlBuilder, IngressUrl};
