+++ crates/rise-backend-docker/src/container_builder.rs
@@ -10,130 +10,25 @@ use std::collections::HashMap;
 use bollard::container::Config;
 use bollard::secret::{EndpointSettings, HostConfig, RestartPolicy, RestartPolicyNameEnum};
 
-use super::labels::{self, BookkeepingLabels, ForwardAuth, TraefikRoute};
+use super::labels::{self, BookkeepingLabels};
 use rise_backend_core::quantity::{
     parse_cpu_millicores, parse_cpu_request_limit, parse_memory_bytes, parse_memory_request_limit,
 };
+use rise_backend_core::traefik_render::TraefikRenderConfig;
 use rise_backend_core::AccessRequirement;
 
-/// Per-route forwardAuth outcome for a single Traefik router.
-enum RouteForwardAuth {
-    /// No auth requirement — emit the router with no forwardAuth middleware.
-    Open,
-    /// Auth required and wired — attach forwardAuth pointing at this address.
-    Gated(String),
-    /// Auth required but no `auth_backend_url` to wire it — withhold this router
-    /// (fail closed) rather than expose an unauthenticated public route.
-    Withheld,
-}
-
-/// One ingress route attached to a routable container.
-#[derive(Debug, Clone)]
-pub struct DesiredRoute {
-    /// Hosts that resolve to this container, priority order.
-    pub hosts: Vec<String>,
-    /// Optional path prefix (`None` / `/` → host-only).
-    pub path_prefix: Option<String>,
-    /// Per-route ingress auth requirement override (`.rise.toml` `[routes].access`).
-    /// `None` means the route inherits the project's access-class requirement; the
-    /// effective requirement decides this router's forwardAuth middleware.
-    pub access: Option<AccessRequirement>,
-}
-
-/// Fully-resolved description of a single container Rise wants running. Built by
-/// the reconciler; consumed by [`build_container`]. All identity fields are
-/// owned strings so the struct can be tested without a live deployment row.
-#[derive(Debug, Clone)]
-pub struct DesiredContainer {
-    pub project: String,
-    /// Name of the project's access class (looked up in
-    /// [`BuilderConfig::access_classes`] to decide whether to stamp Traefik
-    /// forwardAuth middleware labels).
-    pub access_class: String,
-    pub deployment_group: String,
-    pub deployment_id: String,
-    pub deployment_uuid: String,
-    /// Container name within the deployment (`app` for single-container).
-    pub container: String,
-    pub environment: Option<String>,
-    pub image: String,
-    /// `None` for workers (no Service / routing / probe).
-    pub port: Option<u16>,
-    pub cpu: String,
-    pub memory: String,
-    /// Merged env vars as `(KEY, VALUE)` pairs, already in final precedence.
-    pub env: Vec<(String, String)>,
-    /// sha256 of the *entire* merged env (plain + system/RISE_* + secret),
-    /// computed over a deterministically-sorted copy. Drift here forces
-    /// recreation, so editing/deleting any env var of any kind recreates the
-    /// container. See [`super::reconciler::hash_env`].
-    pub env_hash: String,
-    /// Routes for this container (empty for workers / unrouted containers).
-    pub routes: Vec<DesiredRoute>,
-    /// Whether this container should be routable (emit Traefik router/service
-    /// labels). `true` for every infra-bearing deployment of a group — the old
-    /// active and the new Deploying deployment both join the one group-scoped
-    /// Traefik service, and Traefik's per-server health check drains the old
-    /// servers as the new ones come up. `false` only when the router would be
-    /// withheld (unknown access class, or auth required without an
-    /// `auth_backend_url`), so the readiness path doesn't report a never-routed
-    /// container as Healthy.
-    pub routable: bool,
-    /// Recreate-signature hash: sha256 of the fully-rendered Traefik label set
-    /// for this container PLUS whether its app port is published to a loopback
-    /// host port (`publish_app_ports`). Precomputed by the reconciler via
-    /// [`route_hash_for`] so the (pure) diff can compare it against the
-    /// `route-hash` bookkeeping label stamped on the actual container, and force
-    /// a recreate when either changes — a deployment becoming/ceasing to be
-    /// active, or the published-port binding being added/removed. Docker can't
-    /// mutate a running container's labels or port bindings in place, so such a
-    /// transition must be reconciled by recreation.
-    pub route_hash: String,
-    /// Resolved monotonic generation for this container's `--name` suffix
-    /// (`..._g{n}`). NOT an identity field and NOT fed into routing or any hash.
-    /// `compute_desired_for_deployment` can't know it (it depends on the live
-    /// container's current generation), so it seeds `1` as a placeholder and the
-    /// diff resolves the real value before apply: a brand-new slot → `1`, a
-    /// recreate → live `g{n}` + 1.
-    pub generation: u32,
-    /// Zero-based replica index of this container within its spec (`0..N`). IS an
-    /// identity field (folded into [`identity_key`] and the `..._r{n}` name
-    /// segment) so each replica is matched/recreated independently. Deliberately
-    /// NOT fed into routing/recreate hashes, the network alias, or the Traefik
-    /// labels: all replicas of a spec carry one replica-free alias (Docker DNS
-    /// round-robins) and one router+service (Traefik load-balances).
-    pub replica: u32,
-    /// Effective HTTP health-probe path for this container (already resolved from
-    /// the spec's `health_check`): `Some("/healthz")` to probe, or `None` when the
-    /// probe is disabled or the container is a port-less worker. Drives the
-    /// Traefik load-balancer health-check labels (and is thus folded into the
-    /// rendered label set the route-hash covers, so changing the effective health
-    /// path forces a recreate), and gates the rolling-recreate throttle's
-    /// requirement that every OTHER replica be healthy before a running drifted
-    /// replica is recreated. NOT an identity field.
-    pub health_path: Option<String>,
-    /// Traefik load-balancer health-check interval, in seconds (the spec's
-    /// `period_seconds` when set). Only consulted when `health_path` is `Some`, to
-    /// render `...loadbalancer.healthcheck.interval`. `None` → a sensible default.
-    pub health_check_interval_secs: Option<i32>,
-    /// Traefik load-balancer health-check timeout, in seconds (the spec's
-    /// `timeout_seconds` when set). Same gating as `health_check_interval_secs`;
-    /// renders `...loadbalancer.healthcheck.timeout`. `None` → a sensible
-    /// default.
-    pub health_check_timeout_secs: Option<i32>,
-}
-
-/// Default Traefik health-check interval (Go duration `10s`) when the
-/// `health_check` spec sets no `period_seconds`.
-const DEFAULT_HEALTHCHECK_INTERVAL_SECS: i32 = 10;
-/// Default Traefik health-check timeout (Go duration `5s`) when the
-/// `health_check` spec sets no `timeout_seconds`. Matches the Kubernetes
-/// default (`HealthProbeConfig::timeout_seconds` = 5 in
-/// `resource_builder::create_http_probe_with_override`) so the same public
-/// input — a `health_check` with no explicit `timeout_seconds` — yields the
-/// same effective timeout on both backends (no Docker-stricter divergence
-/// that could mark a slow-but-healthy endpoint DOWN on Docker but UP on K8s).
-const DEFAULT_HEALTHCHECK_TIMEOUT_SECS: i32 = 5;
+// The desired-state shape, the name derivations and the Traefik label renderer
+// are backend-agnostic and now live in `rise-backend-core`, shared with the ECS
+// backend. Re-exported here so every existing `container_builder::…` path — in
+// this crate and in `rise-deploy` — resolves unchanged.
+pub use rise_backend_core::desired::{DesiredContainer, DesiredRoute};
+pub use rise_backend_core::naming::{
+    container_name, group_app_name, group_service_base, group_service_name, stable_identity_name,
+};
+pub use rise_backend_core::naming::{group_service_names, MAX_NAME_LEN, MAX_SERVICE_BASE_LEN};
+pub use rise_backend_core::traefik_render::{
+    render_traefik_labels_for, routes_withheld, TraefikRenderConfig as _TraefikRenderConfig,
+};
 
 /// Static controller configuration the builder needs.
 pub struct BuilderConfig<'a> {
@@ -169,116 +64,38 @@ pub struct BuilderConfig<'a> {
     pub publish_app_ports: bool,
 }
 
+impl<'a> BuilderConfig<'a> {
+    /// Project onto the backend-agnostic renderer config. The Docker provider
+    /// needs `traefik.docker.network`, so the network is always `Some` here.
+    pub fn render_config(&self) -> TraefikRenderConfig<'_> {
+        TraefikRenderConfig {
+            label_namespace: self.label_namespace,
+            controller_class: self.controller_class,
+            traefik_entrypoint: self.traefik_entrypoint,
+            traefik_certresolver: self.traefik_certresolver,
+            network: Some(self.traefik_network),
+            auth_backend_url: self.auth_backend_url,
+            access_classes: self.access_classes,
+        }
+    }
+}
+
+/// Compute the `route-hash` recreate signature for a desired container without
+/// building the full create spec, folding in this backend's `publish_app_ports`
+/// binding (a create-time-only property Docker can't mutate in place).
+pub fn route_hash_for(desired: &DesiredContainer, cfg: &BuilderConfig<'_>) -> String {
+    rise_backend_core::traefik_render::route_hash_for(
+        desired,
+        &cfg.render_config(),
+        desired.port.is_some() && cfg.publish_app_ports,
+    )
+}
+
 /// Result of building one container: the deterministic name and the bollard
 /// create config (which already carries the network attachment).
 pub struct BuiltContainer {
     pub name: String,
     pub config: Config<String>,
-}
-
-/// Maximum length of a Docker container name segment we emit before hashing.
-const MAX_NAME_LEN: usize = 63;
-
-/// Compute the live container's `--name`:
-/// `<prefix>_<project>_<group>_<deploymentid>_<container>_r<replica>_g<generation>`,
-/// sanitized to `[a-zA-Z0-9_.-]`, hash-suffixed when longer than
-/// [`MAX_NAME_LEN`]. The `_r{n}_g{n}` suffix is folded into the raw string BEFORE
-/// the length cap, so the >63-char hash-truncation branch still caps the whole
-/// string (suffix included). The replica index keeps each replica's name
-/// distinct; the generation makes a recreated container's name visibly newer
-/// than the one it replaced. Matching is by bookkeeping LABELS (the stable
-/// identity tuple including the replica), never by this name — see
-/// [`group_app_name`] for the group-scoped, deployment-id-free identity used by
-/// DNS / env.
-pub fn container_name(
-    prefix: &str,
-    project: &str,
-    deployment_group: &str,
-    deployment_id: &str,
-    container: &str,
-    replica: u32,
-    generation: u32,
-) -> String {
-    let raw = format!(
-        "{prefix}_{project}_{deployment_group}_{deployment_id}_{container}_r{replica}_g{generation}"
-    );
-    sanitize_and_cap(&raw)
-}
-
-/// Replica- and generation-FREE stable identity name:
-/// `<prefix>_<project>_<group>_<deploymentid>_<container>`, sanitized + capped
-/// the same way as [`container_name`] but with no `_r{n}` / `_g{n}` suffix.
-///
-/// Still deployment-id-BEARING (unlike [`group_app_name`]). Used only to
-/// synthesize a stable per-replica placeholder pod name for diagnostics when a
-/// replica has no live container yet (the diff appends `_r{n}`). The DNS-facing
-/// names — the network alias and the `RISE_CONTAINER_HOST__<NAME>` discovery host
-/// — are GROUP-scoped now (see [`group_app_name`]) so they stay stable across
-/// deployments, not just across replicas/generations of one deployment.
-pub fn stable_identity_name(
-    prefix: &str,
-    project: &str,
-    deployment_group: &str,
-    deployment_id: &str,
-    container: &str,
-) -> String {
-    let raw = format!("{prefix}_{project}_{deployment_group}_{deployment_id}_{container}");
-    sanitize_and_cap(&raw)
-}
-
-/// Group-scoped, deployment-id-FREE application name:
-/// `<prefix>_<project>_<group>_<container>`, sanitized + capped exactly like
-/// [`container_name`] but with NO deployment-id, replica, or generation segment.
-///
-/// This is the stable, deployment-id-free name shared by ALL of a group's
-/// deployments and replicas. EVERY container that belongs to a (project, group,
-/// container) — regardless of which deployment created it — attaches this same
-/// name as its Docker NETWORK ALIAS, so Docker's embedded DNS ROUND-ROBINS the
-/// alias across whatever containers currently carry it (matching the Kubernetes
-/// group Service, whose name is likewise deployment-id-free). It is also the
-/// `RISE_CONTAINER_HOST__<NAME>` sibling-discovery host. Because the name is
-/// stable across deployments, an old and a new deployment of the same group can
-/// share one DNS name during a rolling overlap — foundational for
-/// health-driven rolling-overlap routing.
-pub fn group_app_name(
-    prefix: &str,
-    project: &str,
-    deployment_group: &str,
-    container: &str,
-) -> String {
-    let raw = format!("{prefix}_{project}_{deployment_group}_{container}");
-    sanitize_and_cap(&raw)
-}
-
-/// Sanitize a raw name to `[a-zA-Z0-9_.-]` and hash-truncate it when it exceeds
-/// [`MAX_NAME_LEN`]. Shared by [`container_name`], [`stable_identity_name`] and
-/// [`group_app_name`] so the 63-char cap logic lives in one place.
-/// Deterministic: same input → same output.
-fn sanitize_and_cap(raw: &str) -> String {
-    let sanitized = sanitize_name(raw);
-    if sanitized.len() <= MAX_NAME_LEN {
-        return sanitized;
-    }
-    // Hash the full sanitized name and truncate the prefix to leave room for a
-    // short stable suffix. Deterministic: same inputs → same name.
-    use sha2::{Digest, Sha256};
-    let digest = Sha256::digest(sanitized.as_bytes());
-    let suffix: String = digest.iter().map(|b| format!("{b:02x}")).collect();
-    let suffix = &suffix[..10];
-    let keep = MAX_NAME_LEN - suffix.len() - 1;
-    format!("{}_{}", &sanitized[..keep], suffix)
-}
-
-fn sanitize_name(raw: &str) -> String {
-    raw.chars()
-        .map(|c| {
-            if c.is_ascii_alphanumeric() || c == '_' || c == '.' || c == '-' {
-                c
-            } else {
-                '_'
-            }
-        })
-        .collect()
 }
 
 /// Build the bollard create spec for one desired container.
@@ -299,7 +116,7 @@ pub fn build_container(desired: &DesiredContainer, cfg: &BuilderConfig<'_>) -> B
     // detect create-time-only changes Docker can't apply to a running container
     // in place — routing transitions (a deployment becoming/ceasing to be
     // active) AND whether the app port is published to a loopback host port.
-    let traefik_labels = render_traefik_labels_for(desired, cfg);
+    let traefik_labels = render_traefik_labels_for(desired, &cfg.render_config());
     let route_hash = labels::hash_recreate_signature(
         &traefik_labels,
         desired.port.is_some() && cfg.publish_app_ports,
@@ -477,356 +294,6 @@ pub fn build_container(desired: &DesiredContainer, cfg: &BuilderConfig<'_>) -> B
     };
 
     BuiltContainer { name, config }
-}
-
-/// Maximum length of the human-readable (sanitized) portion of a Traefik
-/// router/service base name, before the `-{hex16}` injective suffix is appended.
-/// Comfortably inside Traefik/Docker name limits even with the suffix and any
-/// per-route `-{idx}` / `-auth` decorations [`group_service_name`] /
-/// [`render_traefik_labels_for`] add on top.
-const MAX_SERVICE_BASE_LEN: usize = 48;
-
-/// Group-scoped Traefik router/service BASE name for a (project, group,
-/// container), deployment-id-FREE so ALL routers/services/middlewares of a
-/// (project, group, container) are named IDENTICALLY across every deployment of
-/// the group. This lets an old and a new deployment share one Traefik service
-/// (their replica containers register as servers of the same load balancer),
-/// setting up health-driven rolling overlap. Mirrors the K8s group
-/// Service/Ingress naming, which is likewise deployment-id-free.
-///
-/// **Injective.** [`sanitize_router_name`] is lossy — it lowercases and
-/// collapses every run of non-`[a-z0-9]` to a single `-`, so distinct tuples
-/// that differ only in separators or case (e.g. `("a.b", …)` vs `("a-b", …)`,
-/// or `Foo` vs `foo`) would otherwise collapse to the SAME router/service name.
-/// Two projects pooled behind one Traefik load balancer is a multi-tenant
-/// boundary break (one project's labels could suppress the other's forwardAuth).
-/// To guarantee distinct tuples get distinct names, we append a short stable
-/// hash of the STRUCTURED `(project, group, container)` tuple — the first 8 bytes
-/// (16 hex chars) of a SHA-256 over a NUL-separated (collision-free, since NUL can't
-/// appear in any field) encoding of the three fields. The human-readable base
-/// is length-capped first so the total stays well within name limits.
-///
-/// Shared by [`render_traefik_labels_for`] (which stamps the labels) and the
-/// reconciler's `serverStatus` lookup (which queries `{service}@docker`) so the
-/// two can't drift on how the service is named.
-pub fn group_service_base(project: &str, deployment_group: &str, container: &str) -> String {
-    use sha2::{Digest, Sha256};
-    let mut sanitized =
-        labels::sanitize_router_name(&format!("{project}-{deployment_group}-{container}"));
-    sanitized.truncate(MAX_SERVICE_BASE_LEN);
-    // Trim any trailing `-` left by truncation so the base reads cleanly before
-    // the suffix join (and never yields a `--` run).
-    let base = sanitized.trim_end_matches('-');
-
-    // Hash the STRUCTURED tuple (not the lossy sanitized string) with NUL field
-    // separators, so two tuples collide here only if they are byte-identical.
-    // 64 bits of suffix (16 hex chars): this name is a multi-tenant boundary and
-    // an attacker controls their own project/group/container names, so a shorter
-    // suffix would be grindable into a deliberate collision with a victim's
-    // service; 64 bits makes that infeasible.
-    let mut hasher = Sha256::new();
-    hasher.update(project.as_bytes());
-    hasher.update([0u8]);
-    hasher.update(deployment_group.as_bytes());
-    hasher.update([0u8]);
-    hasher.update(container.as_bytes());
-    let digest = hasher.finalize();
-    let suffix: String = digest.iter().take(8).map(|b| format!("{b:02x}")).collect();
-
-    if base.is_empty() {
-        suffix
-    } else {
-        format!("{base}-{suffix}")
-    }
-}
-
-/// Per-route Traefik service/router name for route `route_idx` of a container
-/// that emits `route_count` routes. A single-route container uses the bare
-/// [`group_service_base`]; a multi-route container gets per-route services
-/// `{base}-{idx}` (so multiple path prefixes don't collide on one router/service
-/// name). The reconciler queries these SAME names for `serverStatus`, so the
-/// derivation lives in ONE place. Routes are sorted longest-path-prefix-first by
-/// [`render_traefik_labels_for`] before indexing, so callers that need the names
-/// in label order must sort the same way.
-pub fn group_service_name(base: &str, route_idx: usize, route_count: usize) -> String {
-    if route_count > 1 {
-        format!("{base}-{route_idx}")
-    } else {
-        base.to_string()
-    }
-}
-
-/// The full set of Traefik service names a container's routable routes emit, in
-/// the same order [`render_traefik_labels_for`] stamps them (longest path-prefix
-/// first). Empty when the container is not routable / portless / has no host.
-///
-/// The reconciler reads `serverStatus` for the same service(s) via the runtime
-/// [`DesiredContainer`]-free path
-/// ([`super::reconciler::service_names_for_spec`]); this `DesiredContainer`-based
-/// form is the reference used to assert the two derivations agree.
-#[cfg(test)]
-pub fn group_service_names(desired: &DesiredContainer) -> Vec<String> {
-    if desired.port.filter(|_| desired.routable).is_none() {
-        return Vec::new();
-    }
-    let base = group_service_base(
-        &desired.project,
-        &desired.deployment_group,
-        &desired.container,
-    );
-    let mut routes = desired.routes.clone();
-    routes.sort_by(|a, b| {
-        let al = a.path_prefix.as_deref().unwrap_or("/").len();
-        let bl = b.path_prefix.as_deref().unwrap_or("/").len();
-        bl.cmp(&al)
-    });
-    // Enumerate over the FULL sorted list (matching `render_traefik_labels_for`,
-    // where the `{base}-{idx}` index is the position in the full list and the
-    // multi-route test is `routes.len() > 1`), but only emit a name for routes
-    // that actually produce a router — those with at least one host.
-    let route_count = routes.len();
-    routes
-        .iter()
-        .enumerate()
-        .filter(|(_, r)| !r.hosts.is_empty())
-        .map(|(idx, _)| group_service_name(&base, idx, route_count))
-        .collect()
-}
-
-/// Whether all Traefik routers must be withheld for a container because its
-/// access class is unknown, or at least one effective route needs forwardAuth
-/// but no backend URL is configured. Dropping only the gated router would let a
-/// broader public sibling serve that path, so the whole container fails closed.
-///
-/// This is the authoritative predicate for label rendering, desired routability,
-/// and readiness. Route overrides are included so all three paths agree when a
-/// public project tightens only one route.
-pub(crate) fn routes_withheld<'a>(
-    access_class: &str,
-    access_classes: &HashMap<String, AccessRequirement>,
-    auth_backend_url: &str,
-    route_overrides: impl IntoIterator<Item = Option<&'a AccessRequirement>>,
-) -> bool {
-    let mut route_overrides = route_overrides.into_iter().peekable();
-    // A removed/renamed class is a control-plane configuration error. Route
-    // overrides must never weaken that failure into public access: Kubernetes
-    // emits no ingress for the same state, so Docker withholds every router too.
-    let Some(project_requirement) = access_classes.get(access_class) else {
-        return route_overrides.peek().is_some();
-    };
-    if !auth_backend_url.trim().is_empty() {
-        return false;
-    }
-    route_overrides.any(|route_override| {
-        !matches!(
-            rise_backend_core::effective_access_requirement(route_override, project_requirement),
-            AccessRequirement::None
-        )
-    })
-}
-
-/// Render the full Traefik label map for a desired container.
-///
-/// Empty when the container is not routable, has no port (worker), or has no
-/// host to route. Every infra-bearing deployment of a group is routable — the
-/// old active and the new Deploying deployment both advertise a router on the
-/// shared `Host(...)` rule and join the one group-scoped Traefik service, and
-/// Traefik's per-server health check drains the old servers as the new ones come
-/// up (the rolling overlap). `routable` is `false` only when the router is
-/// withheld (unknown access class, or auth required without an
-/// `auth_backend_url`), so a misconfigured deployment never advertises an
-/// unauthenticated router.
-fn render_traefik_labels_for(
-    desired: &DesiredContainer,
-    cfg: &BuilderConfig<'_>,
-) -> HashMap<String, String> {
-    let mut out = HashMap::new();
-    let Some(port) = desired.port.filter(|_| desired.routable) else {
-        return out;
-    };
-
-    // Resolve the project's access requirement. For Authenticated/Member, and
-    // when an internal backend URL is configured, build the Traefik forwardAuth
-    // address pointing at Rise's ingress-auth endpoint. `signin_redirect=1` puts
-    // the handler in Traefik mode (302 to the login page on unauthenticated,
-    // since Traefik has no nginx-style auth-signin). The project name is
-    // URL-encoded into the query. `None` leaves forward_auth unset → no auth.
-    //
-    // FAIL CLOSED on an unknown access class: a project referencing a
-    // missing/removed class must NOT silently become public (mirroring the K8s
-    // path, which errors on an unknown class). Route overrides cannot weaken a
-    // missing class, so suppress every router rather than inventing a default.
-    let requirement = match cfg.access_classes.get(&desired.access_class) {
-        Some(req) => req.clone(),
-        None => {
-            tracing::error!(
-                project = %desired.project,
-                access_class = %desired.access_class,
-                "Access class not configured — withholding all Traefik routers \
-                 to avoid a silent public route"
-            );
-            return out;
-        }
-    };
-    // Per-route effective requirement decides that router's forwardAuth. The
-    // address stamps `&access=<req>` so the shared `ingress_auth` handler enforces
-    // exactly this route group's requirement (never re-matching the request path).
-    // `signin_redirect=1` puts the handler in Traefik mode (302 to login on
-    // unauthenticated). A `None` route gets no middleware (open); an auth route on
-    // a project whose `auth_backend_url` is empty is withheld per-route below.
-    let route_forward_auth = |route_requirement: &AccessRequirement| -> RouteForwardAuth {
-        match route_requirement {
-            AccessRequirement::None => RouteForwardAuth::Open,
-            AccessRequirement::Authenticated | AccessRequirement::Member => {
-                if cfg.auth_backend_url.trim().is_empty() {
-                    RouteForwardAuth::Withheld
-                } else {
-                    RouteForwardAuth::Gated(format!(
-                        "{}/api/v1/auth/ingress?project={}&access={}&signin_redirect=1",
-                        cfg.auth_backend_url.trim().trim_end_matches('/'),
-                        urlencoding::encode(&desired.project),
-                        route_requirement.as_query_param(),
-                    ))
-                }
-            }
-        }
-    };
-
-    // FAIL CLOSED: an unknown access class or an auth-required route without a
-    // forwardAuth address must expose nothing. Stamping a router in either state
-    // could turn a control-plane error into public access.
-    if routes_withheld(
-        &desired.access_class,
-        cfg.access_classes,
-        cfg.auth_backend_url,
-        desired
-            .routes
-            .iter()
-            .filter(|route| !route.hosts.is_empty())
-            .map(|route| route.access.as_ref()),
-    ) {
-        tracing::warn!(
-            project = %desired.project,
-            access_class = %desired.access_class,
-            "Withholding Traefik router(s): access class is unknown or forwardAuth \
-             could not be wired — refusing to expose an unauthenticated public route"
-        );
-        return out;
-    }
-
-    // One router per (host-set × route). Single-container apps have a single
-    // `/` route, so this yields exactly one router. Longest path-prefix first
-    // matches the nginx semantics used by the K8s path.
-    let mut routes = desired.routes.clone();
-    routes.sort_by(|a, b| {
-        let al = a.path_prefix.as_deref().unwrap_or("/").len();
-        let bl = b.path_prefix.as_deref().unwrap_or("/").len();
-        bl.cmp(&al)
-    });
-    // GROUP-scoped router/service base name — deployment-id-FREE, shared with
-    // the reconciler's `serverStatus` lookup via [`group_service_base`] so the
-    // two can't drift. Distinct per-route service names (`{base}-{idx}` when
-    // there is more than one route) keep multiple path prefixes from colliding;
-    // [`group_service_name`] is the single source for that derivation.
-    let route_count = routes.len();
-    let base = group_service_base(
-        &desired.project,
-        &desired.deployment_group,
-        &desired.container,
-    );
-    for (idx, route) in routes.iter().enumerate() {
-        if route.hosts.is_empty() {
-            continue;
-        }
-        // Effective requirement: the route's `access` override, else the
-        // project's (already failed-closed to Member on an unknown class).
-        let route_requirement = route.access.clone().unwrap_or_else(|| requirement.clone());
-        let forward_auth_address = match route_forward_auth(&route_requirement) {
-            RouteForwardAuth::Open => None,
-            RouteForwardAuth::Gated(address) => Some(address),
-            // Unreachable: the pre-loop scan above withholds the whole container
-            // if any route is Withheld. Fail closed defensively regardless.
-            RouteForwardAuth::Withheld => continue,
-        };
-        let router_name = group_service_name(&base, idx, route_count);
-        let traefik = labels::render_traefik_labels(&TraefikRoute {
-            router_name: &router_name,
-            hosts: &route.hosts,
-            path_prefix: route.path_prefix.as_deref(),
-            port,
-            entrypoint: cfg.traefik_entrypoint,
-            network: cfg.traefik_network,
-            certresolver: cfg.traefik_certresolver,
-            forward_auth: forward_auth_address.as_deref().map(|address| ForwardAuth {
-                address,
-                auth_response_headers: "X-Auth-Request-Email,X-Auth-Request-User",
-            }),
-        });
-        out.extend(traefik);
-
-        // Explicit longest-prefix-first router priority (parity with nginx's
-        // implicit longest-match). Traefik would otherwise fall back to implicit
-        // rule-LENGTH priority, which conflates host-rule length with path
-        // specificity; deriving the priority from the route's path-prefix length
-        // makes a more-specific prefix (`/api/v1`) deterministically outrank a
-        // shorter/host-only one (`/`). The `+1` keeps even a host-only (empty/`/`)
-        // route at a positive, non-zero priority.
-        let path_len = route
-            .path_prefix
-            .as_deref()
-            .filter(|p| !p.is_empty() && *p != "/")
-            .map(str::len)
-            .unwrap_or(0);
-        out.insert(
-            format!("traefik.http.routers.{router_name}.priority"),
-            (path_len + 1).to_string(),
-        );
-
-        // Traefik load-balancer health-check labels — emitted when this container
-        // has an effective health path (a `health_check` is configured). A
-        // ready-when-running container (no `health_check`) emits none. The service
-        // name matches the router/service base (`router_name`), so the check
-        // attaches to this route's service. These labels are part of the rendered
-        // set, so toggling them changes the route-hash and forces a recreate
-        // automatically.
-        if let Some(health_path) = desired.health_path.as_deref() {
-            let interval = desired
-                .health_check_interval_secs
-                .unwrap_or(DEFAULT_HEALTHCHECK_INTERVAL_SECS);
-            let timeout = desired
-                .health_check_timeout_secs
-                .unwrap_or(DEFAULT_HEALTHCHECK_TIMEOUT_SECS);
-            out.insert(
-                format!("traefik.http.services.{router_name}.loadbalancer.healthcheck.path"),
-                health_path.to_string(),
-            );
-            out.insert(
-                format!("traefik.http.services.{router_name}.loadbalancer.healthcheck.interval"),
-                format!("{interval}s"),
-            );
-            out.insert(
-                format!("traefik.http.services.{router_name}.loadbalancer.healthcheck.timeout"),
-                format!("{timeout}s"),
-            );
-            out.insert(
-                format!("traefik.http.services.{router_name}.loadbalancer.healthcheck.scheme"),
-                "http".to_string(),
-            );
-        }
-    }
-    out
-}
-
-/// Compute the `route-hash` recreate signature for a desired container without
-/// building the full create spec. Used by the reconciler's diff so a routing
-/// transition (a deployment becoming or ceasing to be active) OR a change in
-/// whether the app port is published to a loopback host port forces a recreate.
-/// Must stay consistent with the hash stamped by [`build_container`].
-pub fn route_hash_for(desired: &DesiredContainer, cfg: &BuilderConfig<'_>) -> String {
-    labels::hash_recreate_signature(
-        &render_traefik_labels_for(desired, cfg),
-        desired.port.is_some() && cfg.publish_app_ports,
-    )
 }
 
 #[cfg(test)]
